requirements.txt
asgiref==3.8.1
Django==5.1.5
sqlparse==0.5.3
python-dotenv-1.0.1